add drag and drop VRM files into app
