+++ apps/client/src/app/constants.ts
